renamed component js-files to index.js and changed imports
